refactor: 3단계 피드백 - Validator 클래스를 domain 패키지로 이동
