+++ src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -8,7 +8,7 @@ import kitchenpos.table.domain.repository.OrderTableRepository;
 import kitchenpos.table.domain.repository.TableGroupRepository;
 import kitchenpos.table.dto.TableGroupRequest;
 import kitchenpos.table.dto.TableGroupResponse;
-import kitchenpos.table.validator.TableValidator;
+import kitchenpos.table.domain.TableValidator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
