+++ src/test/java/kitchenpos/menu/validator/MenuValidatorTest.java
@@ -13,6 +13,7 @@ import java.util.Optional;
 import kitchenpos.common.domain.Price;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuValidator;
 import kitchenpos.menu.domain.repository.MenuGroupRepository;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.product.domain.Product;
