+++ src/test/java/kitchenpos/table/validator/TableValidatorTest.java
@@ -5,6 +5,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 
 import kitchenpos.order.domain.repository.OrderRepository;
+import kitchenpos.table.domain.TableValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
